Refactor `kimm.blocks`
