+++ kimm/blocks/__init__.py
@@ -1,9 +1,6 @@
-from kimm.blocks.base_block import apply_activation
-from kimm.blocks.base_block import apply_conv2d_block
-from kimm.blocks.base_block import apply_se_block
-from kimm.blocks.depthwise_separation_block import (
-    apply_depthwise_separation_block,
-)
-from kimm.blocks.inverted_residual_block import apply_inverted_residual_block
-from kimm.blocks.transformer_block import apply_mlp_block
-from kimm.blocks.transformer_block import apply_transformer_block
+from kimm.blocks.conv2d import apply_conv2d_block
+from kimm.blocks.depthwise_separation import apply_depthwise_separation_block
+from kimm.blocks.inverted_residual import apply_inverted_residual_block
+from kimm.blocks.squeeze_and_excitation import apply_se_block
+from kimm.blocks.transformer import apply_mlp_block
+from kimm.blocks.transformer import apply_transformer_block
